+++ include/roq/ready.hpp
@@ -1,0 +1,82 @@
+/* Copyright (c) 2017-2023, Hans Erik Thrane */
+
+/* !!! THIS FILE HAS BEEN AUTO-GENERATED !!! */
+
+#pragma once
+
+#include "roq/compat.hpp"
+
+#include <fmt/compile.h>
+#include <fmt/format.h>
+
+#include "roq/event.hpp"
+#include "roq/name.hpp"
+#include "roq/trace.hpp"
+
+namespace roq {
+
+//! Marks the end of the initial download phase
+struct ROQ_PUBLIC Ready final {};
+
+template <>
+inline constexpr std::string_view get_name<Ready>() {
+  using namespace std::literals;
+  return "ready"sv;
+}
+
+}  // namespace roq
+
+template <>
+struct fmt::formatter<roq::Ready> {
+  template <typename Context>
+  constexpr auto parse(Context &context) {
+    return std::begin(context);
+  }
+  template <typename Context>
+  auto format(roq::Ready const &, Context &context) const {
+    using namespace std::literals;
+    using namespace fmt::literals;
+    return fmt::format_to(context.out(), R"({{}})"_cf);
+  }
+};
+
+template <>
+struct fmt::formatter<roq::Event<roq::Ready>> {
+  template <typename Context>
+  constexpr auto parse(Context &context) {
+    return std::begin(context);
+  }
+  template <typename Context>
+  auto format(roq::Event<roq::Ready> const &event, Context &context) const {
+    using namespace fmt::literals;
+    return fmt::format_to(
+        context.out(),
+        R"({{)"
+        R"(ready={}, )"
+        R"(message_info={})"
+        R"(}})"_cf,
+        event.value,
+        event.message_info);
+  }
+};
+
+template <>
+struct fmt::formatter<roq::Trace<roq::Ready>> {
+  template <typename Context>
+  constexpr auto parse(Context &context) {
+    return std::begin(context);
+  }
+  template <typename Context>
+  auto format(roq::Trace<roq::Ready> const &event, Context &context) const {
+    using namespace std::literals;
+    using namespace fmt::literals;
+    return fmt::format_to(
+        context.out(),
+        R"({{)"
+        R"(ready={}, )"
+        R"(trace_info={})"
+        R"(}})"_cf,
+        event.value,
+        event.trace_info);
+  }
+};